requirements.txt
streamlit
requests
Pillow
ollama